+++ frontend/src/pages/Messages/Messages.tsx
@@ -1,0 +1,10 @@
+import React from "react";
+
+export const Messages = () => {
+
+	return (
+		<div>
+			Messages page
+		</div>
+	);
+};