+++ frontend/src/types/message/conversation.dto.ts
@@ -1,0 +1,6 @@
+import { MessageDto } from "./message.dto";
+
+export interface ConversationDto {
+  id: number;
+  messages?: MessageDto[]
+}